Add SettingActivity, activity_main.xml
