+++ app/src/main/res/layout/activity_linear.xml
@@ -4,22 +4,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical">
-    <RadioGroup
-        android:id="@+id/radioButtons"
-        android:layout_width="match_parent"
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/button_settings"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_weight="1">
-    <com.google.android.material.radiobutton.MaterialRadioButton
-        android:id="@+id/radioButtonAppThemeLight"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/material_light" />
-    <com.google.android.material.radiobutton.MaterialRadioButton
-        android:id="@+id/radioButtonMyDarkTheme"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/material_dark" />
-   </RadioGroup>
+        android:layout_weight="0.4"
+        android:text="@string/btn_settings"
+        style="@style/CalculatorButtons"/>
 
 
     <Space
